Split k8s coremodel bits into crd subdir
