+++ internal/components/datasource/schema.go
@@ -2,22 +2,10 @@ package datasource
 
 import (
 	"embed"
-	"reflect"
+	"path/filepath"
 
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-
+	"github.com/grafana/grafana/internal/cuectx"
 	"github.com/grafana/thema"
-)
-
-const (
-	cuePath     = "internal/components/datasource"
-	cueFilename = "lineage.cue"
-	schemaName  = "datasource"
-	groupName   = "datasource.core.grafana"
-	// TODO this should be derived from the Thema synv.
-	groupVersion = "v1alpha1"
 )
 
 var (
@@ -25,101 +13,21 @@ var (
 	cueFS embed.FS
 
 	// TODO: this should be generated by Thema.
-	schemaVersion = thema.SV(0, 0)
-
-	// This schema is used for registering CRDs.
-	//
-	// TODO: this should be generated by Thema.
-	schemaOpenapi = apiextensionsv1.JSONSchemaProps{
-		Type: "object",
-		Properties: map[string]apiextensionsv1.JSONSchemaProps{
-			"spec": {
-				Type: "object",
-				Properties: map[string]apiextensionsv1.JSONSchemaProps{
-					"type": {
-						Type: "string",
-					},
-					"typeLogoUrl": {
-						Type: "string",
-					},
-					"access": {
-						Type: "string",
-					},
-					"url": {
-						Type: "string",
-					},
-					"password": {
-						Type: "string",
-					},
-					"user": {
-						Type: "string",
-					},
-					"database": {
-						Type: "string",
-					},
-					"basicAuth": {
-						Type: "boolean",
-					},
-					"basicAuthUser": {
-						Type: "string",
-					},
-					"basicAuthPassword": {
-						Type: "string",
-					},
-					"withCredentials": {
-						Type: "boolean",
-					},
-					"isDefault": {
-						Type: "boolean",
-					},
-					"jsonData": {
-						Type: "string",
-					},
-					"version": {
-						Type: "integer",
-					},
-					"readOnly": {
-						Type: "boolean",
-					},
-				},
-			},
-		},
-	}
+	currentVersion = thema.SV(0, 0)
 )
 
-// Datasource is the Kubernetes-compatible Datasource definition.
-//
-// TODO: this should be generated by Thema.
-type Datasource struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Spec   DatasourceSpec   `json:"spec,omitempty"`
-	Status DatasourceStatus `json:"status,omitempty"`
+// Lineage returns the Thema lineage representing Grafana datasources. The
+// lineage is the canonical specification of the datasource schema, all past
+// schema versions, and mappings that allow migration between schema versions.
+func Lineage(lib thema.Library, opts ...thema.BindOption) (thema.Lineage, error) {
+	return cuectx.LoadGrafanaInstancesWithThema(filepath.Join("internal", "components", "datasource"), cueFS, lib, opts...)
 }
 
-// DeepCopyObject returns a deep copy of Datasource.
+// Model is the canonical Go representation of the current Thema schema version
+// of a Grafana datasource object.
 //
 // TODO: this should be generated by Thema.
-func (in *Datasource) DeepCopyObject() runtime.Object {
-	val := reflect.ValueOf(in).Elem()
-
-	cpy := reflect.New(val.Type())
-	cpy.Elem().Set(val)
-
-	ret, ok := cpy.Interface().(runtime.Object)
-	if !ok {
-		return nil
-	}
-
-	return ret
-}
-
-// DatasourceSpec is the specification of a datasource.
-// This specification is derived from the definition in Thema.
-//
-// TODO: this should be generated by Thema.
-type DatasourceSpec struct {
+type Model struct {
 	Type              string `json:"type"`
 	TypeLogoUrl       string `json:"typeLogoUrl"`
 	Access            string `json:"access"`
@@ -135,35 +43,4 @@ type DatasourceSpec struct {
 	JsonData          string `json:"jsonData,omitempty"`
 	Version           int    `json:"version"`
 	ReadOnly          bool   `json:"readOnly"`
-}
-
-// DatasourceStatus is the status of a datasource (i.e. all API-controller fields go here).
-//
-// TODO: this should be generated by Thema.
-type DatasourceStatus struct{}
-
-// DatasourceList is a Kubernetes-compatible list of Datasource items.
-//
-// TODO: this should be generated by Thema.
-type DatasourceList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []Datasource `json:"items"`
-}
-
-// DeepCopyObject returns a deep copy of DatasourceList.
-//
-// TODO: this should be generated by Thema.
-func (in *DatasourceList) DeepCopyObject() runtime.Object {
-	val := reflect.ValueOf(in).Elem()
-
-	cpy := reflect.New(val.Type())
-	cpy.Elem().Set(val)
-
-	ret, ok := cpy.Interface().(runtime.Object)
-	if !ok {
-		return nil
-	}
-
-	return ret
 }
