+++ pkg/schema/loader.go
@@ -3,6 +3,7 @@ package schema
 import (
 	"context"
 	"errors"
+	"fmt"
 	"io"
 	"io/fs"
 	"path/filepath"
@@ -107,7 +108,7 @@ func (l *ThemaSchemaLoader) LoadSchema(ctx context.Context, opts ThemaLoaderOpts
 
 	zsch, err := lin.Schema(opts.SchemaVersion)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("no schema with version %v in lineage %s: %w", opts.SchemaVersion, lin.Name(), err)
 	}
 
 	if err := thema.AssignableTo(zsch, opts.SchemaType); err != nil {
